Add simple string cache to the change list
The change list contains a lot of recurring strings, such as attribute names. In many cases, this also includes things like class names, which are repeated many times in e.g. list displays.

This patch adds a simple cache for those strings. Whenever a string is encountered, it's checked against the cache. If it doesn't exist in the cache, it's added, and an opcode is emmitted sending that string to JS.

After computing the full change list, the cache is iterated, and all strings not in use in the current change list are evicted, again emitting an opcode per entry instructing the JS side to drop that string, too.

This seems to strike a reasonable balance between keeping the cache across multiple updates—where it helps in scenarios where e.g. multiple list entries are added one after the other—and wasting memory. In particular, right after a string was sent, it's almost certainly in use in the resulting DOM, so keeping it around in the string cache, too, shouldn't waste memory for the string, just for the cache entry itself.

Note that this patch doesn't use the string cache for text nodes: those are deemed less likely to benefit from the cache.