Use the same function for all DOM event handlers
This changes the event handler setup to reuse a single function for all DOM event handling. Instead of creating a new listener for each event, the same function is registered as a handler. It looks up properties on the `this` object to use as values for `a` and `b` passed into the trampoline.

This gets rid of all or almost all GC activity in a simple handler-heavy test case I looked at. AFAICT it should be entirely neutral regarding functionality.