Automatically create "templates" for `Cached<R>`
This adds a `R: Default` trait bound to `Cached<R>`, and the first time any
instance of `Cached<R>` is rendered, we first render the default `R` instance
and save it as a "template" for all future `Cached<R>` renders. The subsequent
renders will call `cloneNode` (with the "deep" parameter set to true) on the
template and then diff against the cloned node, rather than creating the whole
subtree from scratch. This cuts down on DOM method invocations and traversal
instructions.