Optimize emitted DOM traversal instructions
This commit introduces a traversal path optimizer. The goal is that while we
have to walk over large virtual DOM trees when diffing, if we don't end up
actually emitting any change list instructions that mutate the physical DOM,
then we don't emit the traversal instructions either. We do this by keeping
track of where the last location we "committed" the traversal in order to mutate
something, and if we go down into children and back up to the parent without
commiting again, then we optimize that traversal into a no-op. We only emit the
optimized traversal path once we "commit" the traversal.